pyproject.toml
[project]
name = "wavesim"
version = "0.2.0a1"
description = "A library for simulating wave propagation using the modified Born series"
authors = [ {name = "Ivo Vellekoop", email = "i.m.vellekoop@utwente.nl"}, 
            {name = "Swapnil Mache", email = "s.mache@utwente.nl"} ]
readme = "README.md"
repository = "https://github.com/IvoVellekoop/wavesim_py"
documentation = "https://wavesim.readthedocs.io/en/latest/"
classifiers = [
    'Programming Language :: Python :: 3',
    'Operating System :: OS Independent',
]
requires-python = ">=3.11,<3.13"
dependencies = [
    "numpy<2.0.0",
    "matplotlib>=3.9.1",
    "scipy>=1.14.0",
    "porespy",
    "scikit-image<0.23",
    "torch"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.poetry.dependencies]
# no explicit numpy dependency is stated so that pip can figure out the version to install
# based on the requirements of cupy, scipy, and numba
python = "^3.11"
scipy = "^1.15.0"
numba = ">=0.60.0"
cupy-cuda12x = "^13.4.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = ">=4.1.2"
sphinx_mdinclude = ">= 0.5.0"
sphinx-rtd-theme = ">= 2.0.0"
sphinx-autodoc-typehints = ">= 1.11.0"
sphinxcontrib-bibtex = ">= 2.6.0"
sphinx-markdown-builder = ">= 0.6.6"
sphinx-gallery = ">= 0.15.0"
